+++ CHANGELOG.md
@@ -128,6 +128,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - scripts: force cd / for all PostgreSQL scripts [PR #1608]
 - dird: remove optimize_for_size/optimize_for_speed [PR #1613]
 - Improve python plugin configuration [PR #1619]
+- restore: add fileregex parameter [PR #1618]
 
 ### Removed
 - cats: remove remains of deprecated databases [PR #1377]
@@ -338,5 +339,6 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1608]: https://github.com/bareos/bareos/pull/1608
 [PR #1611]: https://github.com/bareos/bareos/pull/1611
 [PR #1613]: https://github.com/bareos/bareos/pull/1613
+[PR #1618]: https://github.com/bareos/bareos/pull/1618
 [PR #1619]: https://github.com/bareos/bareos/pull/1619
 [unreleased]: https://github.com/bareos/bareos/tree/master